+++ Tunny/Storage/IStorage.cs
@@ -1,3 +1,5 @@
+using Optuna.Study;
+
 namespace Tunny.Storage
 {
     public interface IStorage : ICreateStorage