+++ CHANGELOG.md
@@ -6,6 +6,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 Please see [here](https://github.com/hrntsm/Tunny/releases) for the data released for each version.
 
+## [v0.10.0] -2024-01-27
+
+### Changed
+
+- Use new JournalStorage handler to improve stability.
+  - old format journal storage result new can open in tunny.
+
+### Fixed
+
+- Does not work when human-in-the-loop continues.
+- Error when reading older versions of JournalStorage.
+- without constraint notification timing when output result.
+- in-memory optimization result saving error.
+
 ## [v0.9.1] -2023-12-30
 
 ### Fixed